修改：分拣码批量匹配

+++ produsupsys-common/src/main/java/com/ruoyi/common/bean/po/PostPpcPrice.java
@@ -9,7 +9,7 @@ public class PostPpcPrice {
     /**
      *  user_id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 菜单: 0:根据数据设备查询 1:匹配查询，2:地址查询，3:面单打印，4:号码呼转 menu
@@ -30,7 +30,7 @@ public class PostPpcPrice {
      *
      * @mbggenerated
      */
-    public PostPpcPrice(Integer id, Integer userId, Integer menu, Integer source, Double price) {
+    public PostPpcPrice(Integer id, Long userId, Integer menu, Integer source, Double price) {
         this.id = id;
         this.userId = userId;
         this.menu = menu;
@@ -47,44 +47,44 @@ public class PostPpcPrice {
     }
 
     /**
-     * 
-    
-     * @return id 
+     *
+
+     * @return id
      */
     public Integer getId() {
         return id;
     }
 
     /**
-     * 
-    
-     * @param id 
+     *
+
+     * @param id
      */
     public void setId(Integer id) {
         this.id = id;
     }
 
     /**
-     * 
-    
-     * @return user_id 
+     *
+
+     * @return user_id
      */
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
     /**
-     * 
-    
-     * @param userId 
+     *
+
+     * @param userId
      */
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
     /**
      * 菜单: 0:根据数据设备查询 1:匹配查询，2:地址查询，3:面单打印，4:号码呼转
-    
+
      * @return menu 菜单: 0:根据数据设备查询 1:匹配查询，2:地址查询，3:面单打印，4:号码呼转
      */
     public Integer getMenu() {
@@ -93,7 +93,7 @@ public class PostPpcPrice {
 
     /**
      * 菜单: 0:根据数据设备查询 1:匹配查询，2:地址查询，3:面单打印，4:号码呼转
-    
+
      * @param menu 菜单: 0:根据数据设备查询 1:匹配查询，2:地址查询，3:面单打印，4:号码呼转
      */
     public void setMenu(Integer menu) {
@@ -102,7 +102,7 @@ public class PostPpcPrice {
 
     /**
      * 数据来源: 0:根据菜单查询 :1WEB，2APP，3接口
-    
+
      * @return source 数据来源: 0:根据菜单查询 :1WEB，2APP，3接口
      */
     public Integer getSource() {
@@ -111,7 +111,7 @@ public class PostPpcPrice {
 
     /**
      * 数据来源: 0:根据菜单查询 :1WEB，2APP，3接口
-    
+
      * @param source 数据来源: 0:根据菜单查询 :1WEB，2APP，3接口
      */
     public void setSource(Integer source) {
@@ -120,7 +120,7 @@ public class PostPpcPrice {
 
     /**
      * 单价:元
-    
+
      * @return price 单价:元
      */
     public Double getPrice() {
@@ -129,7 +129,7 @@ public class PostPpcPrice {
 
     /**
      * 单价:元
-    
+
      * @param price 单价:元
      */
     public void setPrice(Double price) {
